% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{block_idx}
\alias{block_idx}
\title{Get the index range for exposure k's block in a stacked vector (internal)}
\usage{
block_idx(offsets, k)
}
\arguments{
\item{offsets}{Output of compute_offsets()}

\item{k}{Exposure index (1-based)}
}
\value{
Integer vector of indices for exposure k's block
}
\description{
Get the index range for exposure k's block in a stacked vector (internal)
}
\keyword{internal}
