% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compute_offsets}
\alias{compute_offsets}
\title{Compute cumulative block offsets for m stacked exposure blocks (internal)}
\usage{
compute_offsets(nPC_vec)
}
\arguments{
\item{nPC_vec}{Integer vector of length m, number of components per exposure}
}
\value{
Integer vector of length m+1: c(0, cumsum(nPC_vec))
}
\description{
Compute cumulative block offsets for m stacked exposure blocks (internal)
}
\keyword{internal}
