% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_true_effect_function}
\alias{get_true_effect_function}
\title{Get true effect function for simulation}
\usage{
get_true_effect_function(model_code)
}
\arguments{
\item{model_code}{Model code ('0'-'9') specifying the effect shape function}
}
\value{
Function that takes time as input and returns effect value
}
\description{
Get true effect function for simulation
}
\keyword{internal}
