% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_true_shape_values}
\alias{get_true_shape_values}
\title{Get true shape values for simulation}
\usage{
get_true_shape_values(workGrid, XYmodel)
}
\arguments{
\item{workGrid}{Grid of time points for evaluation}

\item{XYmodel}{Model code ('0'-'9') specifying the true effect shape}
}
\value{
Vector of true effect values at workGrid time points
}
\description{
Get true shape values for simulation
}
\keyword{internal}
