% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getY_multi_exposure}
\alias{getY_multi_exposure}
\title{Generate outcome from exposures}
\usage{
getY_multi_exposure(
  RES,
  X1Ymodel = "1",
  X2Ymodel = "1",
  X1_effect = TRUE,
  X2_effect = TRUE,
  outcome_type = "continuous"
)
}
\arguments{
\item{RES}{Output from getX_multi_exposure() or getX_multi_exposure_mediation()}

\item{X1Ymodel}{Effect model for X1 (0-9)}

\item{X2Ymodel}{Effect model for X2 (0-9)}

\item{X1_effect}{Include X1 effect?}

\item{X2_effect}{Include X2 effect?}

\item{outcome_type}{"continuous" or "binary"}
}
\value{
Data frame with outcome Y
}
\description{
Generate outcome from exposures
}
