% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_beta.R
\name{gmm_lm_onesample}
\alias{gmm_lm_onesample}
\title{GMM estimation for continuous outcome}
\usage{
gmm_lm_onesample(X, Y, Z, beta0 = NA)
}
\arguments{
\item{X}{Matrix of exposure principal components (N x K)}

\item{Y}{Outcome vector (length N)}

\item{Z}{Genetic instrument matrix (N x J)}

\item{beta0}{Initial values for beta (default NA, uses zero initialization)}
}
\value{
List with gmm_est, gmm_se, variance_matrix, gmm_pval, Q_stat, Q_pval
}
\description{
GMM estimation for continuous outcome
}
