% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmvmr-package.R
\name{mvfmr-package}
\alias{mvfmr-package}
\alias{_PACKAGE}
\title{mvfmr: Multivariable Functional Mendelian Randomization}
\description{
Implements Multivariable Functional Mendelian randomization to estimate
time-varying causal effects of multiple correlated longitudinal exposures.
}
\author{
Nicole Fontana
}
