% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recycle_arg}
\alias{recycle_arg}
\title{Recycle a per-exposure scalar argument to length m (internal)}
\usage{
recycle_arg(x, m, default = NA)
}
\arguments{
\item{x}{NULL, a single value, or a vector of length m}

\item{m}{Number of exposures}

\item{default}{Value used to fill each of the m entries when x is NULL}
}
\value{
Vector of length m
}
\description{
Recycle a per-exposure scalar argument to length m (internal)
}
\keyword{internal}
