% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getX_multi_exposure}
\alias{getX_multi_exposure}
\title{Generate multi-exposure data with genetic instruments}
\usage{
getX_multi_exposure(
  N = 10000,
  J = 30,
  ZXmodel = "A",
  nSparse = 10,
  NT = 1000,
  TT = 50,
  n_exposures = 2,
  shared_effect = TRUE,
  separate_G = FALSE,
  shared_G_proportion = 0.15
)
}
\arguments{
\item{N}{Sample size}

\item{J}{Number of genetic instruments (per exposure, if separate_G = TRUE)}

\item{ZXmodel}{Model type (currently not used)}

\item{nSparse}{Number of sparse observations per subject}

\item{NT}{Number of points}

\item{TT}{Max observation period}

\item{n_exposures}{Number of exposures to simulate (m)}

\item{shared_effect}{Whether all exposures share the same time-varying confounding}

\item{separate_G}{Whether to use separate instruments for each exposure}

\item{shared_G_proportion}{Proportion of shared instruments (0-1)}
}
\value{
List with per-exposure sparse data and genetic instruments
}
\description{
Generate multi-exposure data with genetic instruments
}
\examples{
set.seed(1)
sim_data <- getX_multi_exposure(N = 50, J = 8, nSparse = 5, n_exposures = 2)
length(sim_data$exposures)
dim(sim_data$details$G)
}
