% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{getY_multi_exposure}
\alias{getY_multi_exposure}
\title{Generate outcome from exposures}
\usage{
getY_multi_exposure(
  RES,
  XYmodels = NULL,
  X_effects = NULL,
  outcome_type = "continuous"
)
}
\arguments{
\item{RES}{Output from getX_multi_exposure() or getX_multi_exposure_mediation()}

\item{XYmodels}{Length-m vector of effect models per exposure, one of '0'-'9' (default: '1' for all)}

\item{X_effects}{Length-m logical vector: include each exposure's effect? (default: TRUE for all)}

\item{outcome_type}{"continuous" or "binary"}
}
\value{
Data frame with outcome Y
}
\description{
Generate outcome from exposures
}
\examples{
set.seed(1)
sim_data <- getX_multi_exposure(N = 50, J = 8, nSparse = 5, n_exposures = 2)
dat <- getY_multi_exposure(sim_data, XYmodels = c("2", "8"), outcome_type = "continuous")
head(dat$Y)
}
