% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_beta.R
\name{gmm_twosample_simple}
\alias{gmm_twosample_simple}
\title{Two-sample GMM}
\usage{
gmm_twosample_simple(bx, by, sy, ny)
}
\arguments{
\item{bx}{Matrix J x K of first-stage coefficients (SNP -> PC associations)}

\item{by}{Vector length J of outcome GWAS betas}

\item{sy}{Vector length J of outcome GWAS standard errors}

\item{ny}{Outcome GWAS sample size}
}
\value{
List with gmm_est, gmm_se, variance_matrix, gmm_pval, Q_stat, Q_df, Q_pval
}
\description{
Two-sample GMM
}
